bwt_sort - bug fix and cosmetic changes
Changes are cosmetic on IvyBridge, but could be significant on
CPUs with expensive unaligned memory accesses.

+++ 32K_bwt_mtf_rle_c0/bwt_mtf_rle_e.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 
 void bwt_sort(
-  int32_t* dst_tmp, // length = srclen*3
+  uint64_t* dst_tmp, // length = ((srclen*3+256)*sizeof(int32_t))/sizeof(uint64_t)
   uint8_t* src,     // length = srclen+8, characters [srclen..srclen+7] modified, others preserved
   int      srclen);
